test(api): move the detail-endpoint test out of the VCR module
Adversarial review found it vacuous, and it was. test_contracts.py carries a
file-level pytest.mark.vcr, and vcrpy intercepts ahead of ASGITransport — so the
test replayed its own committed cassette and never reached the app or the
database. Proven: with the expiry predicate deleted, the three service tests go
red and this one still passed. The detail/list asymmetry, which is this branch's
headline decision, had zero effective coverage.

Moved to test_contract_filters.py (asyncio mark only, no VCR) and deleted the
cassette. Re-proven the other way: with the predicate removed the relocated test
now fails on `assert 1 == 0`.

Also disarms a landmine in the same module: eight fixture contracts hardcoded
date_expired=2025-01-02, long past. They pass today only because vcrpy replays
recorded responses; the first re-record would have failed them all with a
baffling total==0. Now relative to now.

Adds a 30s lock_timeout to the index migration. It runs as Render's pre-deploy
command while the outgoing instance still serves, and that instance's ingestion
holds an open transaction on `contracts` across its upsert phase; CREATE INDEX
needs a SHARE lock, so an overlapping deploy would otherwise wait indefinitely
and stall the release. Bounded wait turns a stall into a visible, retryable
failure.

Corrects an overstated comment: func.now() guards against application/container
clock skew, not index usability — a bound timezone-aware parameter is equally
indexable.
